fix gcc 10 warnings/errors

+++ src/gui/widgets/bar_slider.c
@@ -108,13 +108,15 @@ bar_slider_draw_cb (
   if (real_val < real_zero)
     {
       cairo_rectangle (
-        cr, val_px, 0, zero_px - val_px, height);
+        cr, (double) val_px, 0,
+        (double) (zero_px - val_px), height);
     }
   /* draw from zero to val */
   else
     {
       cairo_rectangle (
-        cr, zero_px, 0, val_px - zero_px, height);
+        cr, (double) zero_px, 0,
+        (double) (val_px - zero_px), height);
     }
   cairo_fill (cr);
